A huge commit
Redid pretty much the entire ASM system, now uses a simpler (more modular) ASM structure based on one primary transformer and a bytecode helper class.
Fixed some hard-to-find chunk generating/loading issues (surprisingly) by implementing the start of chunk generation threading. The chunk generation/loading system on the server side now uses a futures system, allowing for threads to not conflict when multiple attempt to load the same chunk at once. This also allows for chunk threading in the future, as these futures can be delegated into threads. There is a config in place for this; however, it is not recommended as of now!
Refactored many concurrent classes to use more efficient non-blocking structures (atomics).
Added Spool's icon as a mod icon to be shown in-game (it took me too long to figure out how to do this).
Watchdog interval is now a config option.
And a bunch of other miscellaneous changes that I can't seem to remember here.