A lot of changes!
New load-balancing thread manager! This thread manager, based on a load function decided per-thread, dynamically chooses which threads should get their own thread and which should be left to be remapped to run on the main thread. This is currently used for the dimension threading pool; however, it's toggleable via a config option.

Moved around a lot of thread management and IMC stuff to its own classes.
A lot of miscellaneous changes to a lot of classes to reduce *auto-boxing/unboxing*!

Finally fixed LAN servers not properly deconstructing the distance manager when all players other than the host leave.
Fixed an issue with distance manager interfering across worlds. The distance manager should now be completely world-independent. This means that a player with a distance thread in one dimension will not interfere with a player in another dimension with their own distance thread, regardless of relative position to each-other (inside the dim coordinates).